fixing script resolution in extension (#1449)
* removing duplicate state

* move fix command to self file

* further cleanup

* cleanup in various files

* rename

* fix type

* more cleanup

* clean logic

* no sideeffect in getting project

* ✨ refactor state management and cleanup parsing logic

Removed redundant state parsing logic for improved clarity and efficiency.

* fix gists
