+++ .vscode/settings.json
@@ -128,6 +128,7 @@
         "OPENROUTER",
         "optillm",
         "packagemanagers",
+        "parameterquickpick",
         "Peli",
         "pelikhan",
         "pixtral",
@@ -149,6 +150,7 @@
         "sarif",
         "SARIFF",
         "scriptid",
+        "scriptquickpick",
         "secretscanner",
         "sglang",
         "sidenote",
