+++ packages/core/src/workdir.ts
@@ -6,10 +6,12 @@ import {
     STATS_DIR_NAME,
 } from "./constants"
 import { randomHex } from "./crypto"
+import { genaiscriptDebug } from "./debug"
 import { ensureDir } from "./fs"
 import { gitIgnoreEnsure } from "./gitignore"
 import { host } from "./host"
 import { sanitizeFilename } from "./sanitize"
+const dbg = genaiscriptDebug("dirs")
 
 /**
  * Constructs a resolved file path within the `.genaiscript` directory of the project.
@@ -58,12 +60,14 @@ function createDatedFolder(id: string) {
  * @returns The resolved path for the specified run directory.
  */
 export function getRunDir(scriptId: string, runId: string) {
+    dbg(`run: %s %s`, scriptId, runId)
     const name = createDatedFolder(runId)
     const out = dotGenaiscriptPath(
         RUNS_DIR_NAME,
         host.path.basename(scriptId).replace(GENAI_ANYTS_REGEX, ""),
         name
     )
+    dbg("run dir: %s", out)
     return out
 }
 
@@ -77,12 +81,15 @@ export function getRunDir(scriptId: string, runId: string) {
  *          for the converted files.
  */
 export function getConvertDir(scriptId: string) {
-    const name = createDatedFolder(randomHex(6))
+    const runId = randomHex(6)
+    dbg(`convert: %s %s`, scriptId, runId)
+    const name = createDatedFolder(runId)
     const out = dotGenaiscriptPath(
         CONVERTS_DIR_NAME,
         host.path.basename(scriptId).replace(GENAI_ANYTS_REGEX, ""),
         name
     )
+    dbg("convert dir: %s", out)
     return out
 }
