+++ packages/vscode/src/docsnotebook.ts
@@ -135,7 +135,6 @@ function activateNotebookExecutor(state: ExtensionState) {
                 const parameters = { ...heap, ...vars }
                 await state.requestAI({
                     scriptId: template.id,
-                    template,
                     label: "Executing cell",
                     parameters,
                     fragment,
