+++ packages/vscode/src/statusbar.ts
@@ -13,15 +13,14 @@ export function activateStatusBar(state: ExtensionState) {
     )
     statusBarItem.command = "genaiscript.request.status"
     const updateStatusBar = async () => {
-        const { parsing, aiRequest, languageChatModels, host } = state
+        const { aiRequest, languageChatModels, host } = state
         const { server } = host
         const { status } = server
         const { computing, progress, options } = aiRequest || {}
-        const { template, fragment } = options || {}
+        const { fragment } = options || {}
         const { tokensSoFar } = progress || {}
         statusBarItem.text = toStringList(
             `${
-                parsing ||
                 status === "starting" ||
                 status === "stopping" ||
                 (computing && !tokensSoFar)
@@ -37,9 +36,6 @@ export function activateStatusBar(state: ExtensionState) {
                     ? `server: [${server.authority}](${server.browserUrl})`
                     : `server: ${status}`,
                 fragment?.files?.[0],
-                template
-                    ? `-  script: ${template.title} (${template.id})`
-                    : undefined,
                 ...Object.entries(languageChatModels).map(
                     ([m, c]) => `-  language chat model: ${m} -> ${c}`
                 )
